utils: move nodepool pkg under tests
The package is using another package from `test/...`
and is not being used by any other packages but packages under
`test/...`.

so let's move it also under `test` to be fully compatible with
the project packages layout.
